finalizing
rename a lot of types; fix issue with storage buffers; remove old persistent and pool implementations
