+++ vulkano/src/descriptor_set/mod.rs
@@ -74,30 +74,34 @@
 //!   `DescriptorSet`. It is what you pass to the draw functions.
 
 pub use self::collection::DescriptorSetsCollection;
-pub use self::fixed_size_pool::FixedSizeDescriptorSetsPool;
 use self::layout::DescriptorSetLayout;
 pub use self::persistent::PersistentDescriptorSet;
-pub use self::persistent::PersistentDescriptorSetBuildError;
-pub use self::persistent::PersistentDescriptorSetError;
+pub use self::single_layout_pool::SingleLayoutDescSetPool;
 use self::sys::UnsafeDescriptorSet;
 use crate::buffer::BufferAccess;
 use crate::descriptor_set::layout::DescriptorDescTy;
 use crate::device::DeviceOwned;
+use crate::format::Format;
 use crate::image::view::ImageViewAbstract;
+use crate::image::view::ImageViewType;
+use crate::OomError;
 use crate::SafeDeref;
 use crate::VulkanObject;
 use smallvec::SmallVec;
+use std::error;
+use std::fmt;
 use std::hash::Hash;
 use std::hash::Hasher;
 use std::sync::Arc;
 
+mod builder;
 mod collection;
-pub mod fixed_size_pool;
 pub mod layout;
 pub mod persistent;
 pub mod pool;
+mod resources;
+pub mod single_layout_pool;
 pub mod sys;
-pub mod runtime;
 
 /// Trait for objects that contain a collection of resources that will be accessible by shaders.
 ///
@@ -286,4 +290,178 @@ where
     fn from(descriptor_set: S) -> Self {
         Self::new(descriptor_set, std::iter::empty())
     }
+}
+
+/// Error related to descriptor sets.
+#[derive(Debug, Clone)]
+pub enum DescriptorSetError {
+    /// Builder is already within an array.
+    AlreadyInArray,
+
+    /// Builder is not in an array.
+    NotInArray,
+
+    /// Array doesn't contain the correct amount of descriptors
+    ArrayLengthMismatch {
+        /// Expected length
+        expected: u32,
+        /// Obtained length
+        obtained: u32,
+    },
+
+    /// Builder doesn't expect anymore descriptors
+    TooManyDescriptors,
+
+    /// Runtime arrays must be the last binding in a set.
+    RuntimeArrayMustBeLast,
+
+    /// Not all descriptors have been added.
+    DescriptorsMissing {
+        /// Expected bindings
+        expected: usize,
+        /// Obtained bindings
+        obtained: usize,
+    },
+
+    /// The buffer is missing the correct usage.
+    MissingBufferUsage(MissingBufferUsage),
+
+    /// The image is missing the correct usage.
+    MissingImageUsage(MissingImageUsage),
+
+    /// Expected one type of resource but got another.
+    WrongDescriptorType,
+
+    /// Resource belongs to another device.
+    ResourceWrongDevice,
+
+    /// The format of an image view doesn't match what was expected.
+    ImageViewFormatMismatch {
+        /// Expected format.
+        expected: Format,
+        /// Format of the image view that was passed.
+        obtained: Format,
+    },
+
+    /// The type of an image view doesn't match what was expected.
+    ImageViewTypeMismatch {
+        /// Expected type.
+        expected: ImageViewType,
+        /// Type of the image view that was passed.
+        obtained: ImageViewType,
+    },
+
+    /// Expected a multisampled image, but got a single-sampled image.
+    ExpectedMultisampled,
+
+    /// Expected a single-sampled image, but got a multisampled image.
+    UnexpectedMultisampled,
+
+    /// The number of array layers of an image doesn't match what was expected.
+    ArrayLayersMismatch {
+        /// Number of expected array layers for the image.
+        expected: u32,
+        /// Number of array layers of the image that was added.
+        obtained: u32,
+    },
+
+    /// The image view has a component swizzle that is different from identity.
+    NotIdentitySwizzled,
+
+    /// The image view isn't compatible with the sampler.
+    IncompatibleImageViewSampler,
+
+    /// The builder has previously return an error and is an unknown state.
+    BuilderPoisoned,
+
+    /// Out of memory
+    OomError(OomError),
+
+    /// Operation can not be performed on an empty descriptor.
+    DescriptorIsEmpty,
+
+    /// Expected a non-arrayed image, but got an arrayed image.
+    UnexpectedArrayed,
+
+    /// A required feature is missing from the Device
+    MissingFeature(MissingFeature),
+}
+
+impl From<OomError> for DescriptorSetError {
+    fn from(error: OomError) -> Self {
+        Self::OomError(error)
+    }
+}
+
+impl error::Error for DescriptorSetError {}
+
+impl fmt::Display for DescriptorSetError {
+    #[inline]
+    fn fmt(&self, fmt: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+        write!(
+            fmt,
+            "{}",
+            match *self {
+                Self::AlreadyInArray => "builder is already within an array",
+                Self::NotInArray => "builder is not in an array",
+                Self::ArrayLengthMismatch { .. } =>
+                    "array doesn't contain the correct amount of descriptors",
+                Self::TooManyDescriptors => "builder doesn't expect anymore descriptors",
+                Self::RuntimeArrayMustBeLast => "runtime arrays must be the last binding in a set",
+                Self::DescriptorsMissing { .. } => "not all descriptors have been added",
+                Self::MissingBufferUsage(_) => "the buffer is missing the correct usage",
+                Self::MissingImageUsage(_) => "the image is missing the correct usage",
+                Self::WrongDescriptorType => "expected one type of resource but got another",
+                Self::ResourceWrongDevice => "resource belongs to another device",
+                Self::ImageViewFormatMismatch { .. } =>
+                    "the format of an image view doesn't match what was expected",
+                Self::ImageViewTypeMismatch { .. } =>
+                    "the type of an image view doesn't match what was expected",
+                Self::ExpectedMultisampled =>
+                    "expected a multisampled image, but got a single-sampled image",
+                Self::UnexpectedMultisampled =>
+                    "expected a single-sampled image, but got a multisampled image",
+                Self::ArrayLayersMismatch { .. } =>
+                    "the number of array layers of an image doesn't match what was expected",
+                Self::NotIdentitySwizzled =>
+                    "the image view has a component swizzle that is different from identity",
+                Self::IncompatibleImageViewSampler =>
+                    "the image view isn't compatible with the sampler",
+                Self::BuilderPoisoned =>
+                    "the builder has previously return an error and is an unknown state",
+                Self::OomError(_) => "out of memory",
+                Self::DescriptorIsEmpty => "operation can not be performed on an empty descriptor",
+                Self::UnexpectedArrayed => "expected a non-arrayed image, but got an arrayed image",
+                Self::MissingFeature(_) => "a required feature is missing from the Device",
+            }
+        )
+    }
+}
+
+// Part of the DescriptorSetError for the case
+// of missing usage on a buffer.
+#[derive(Debug, Clone)]
+pub enum MissingBufferUsage {
+    StorageBuffer,
+    UniformBuffer,
+    StorageTexelBuffer,
+    UniformTexelBuffer,
+}
+
+// Part of the DescriptorSetError for the case
+// of missing usage on an image.
+#[derive(Debug, Clone)]
+pub enum MissingImageUsage {
+    InputAttachment,
+    Sampled,
+    Storage,
+}
+
+// Part of the DescriptorSetError for the case
+// of missing features on a device.
+#[derive(Debug, Clone)]
+pub enum MissingFeature {
+    RuntimeDescriptorArray,
+    DescriptorBindingVariableDescriptorCount,
+    DescriptorBindingPartiallyBound,
 }
